refactor(web): move directory of modals to `shared`
