+++ apps/web/src/shared/constants/delete.ts
