+++ apps/web/src/shared/constants/index.ts
@@ -5,3 +5,4 @@ export * from './day';
 export * from './board';
 export * from './location';
 export * from './profile';
+export * from './delete';
