+++ apps/web/src/shared/types/delete.ts
@@ -1,3 +1,3 @@
-import { DELETE_MODAL } from '../model';
+import { DELETE_MODAL } from '../constants';
 
 export type DeleteModalItem = keyof typeof DELETE_MODAL;
