+++ apps/web/src/shared/types/index.ts
@@ -2,3 +2,4 @@ export * from './modal';
 export * from './board';
 export * from './location';
 export * from './profile';
+export * from './delete';
