--- apps/web/src/widgets/modal/types/index.ts
@@ -1,4 +1,0 @@
-export * from './post';
-export * from './vote';
-export * from './delete';
-export * from './schedule';
