--- apps/web/src/widgets/modal/ui/index.ts
@@ -1,11 +1,0 @@
-export { default as LogoutModal } from './LogoutModal';
-export { default as AlarmModal } from './AlarmModal';
-export { default as DeleteModal } from './DeleteModal';
-export { default as UpdateModal } from './UpdateModal';
-export { default as CreateGroupBoardModal } from './CreateGroupBoardModal';
-
-export * from './create-post';
-export * from './create-schedule';
-export * from './project';
-export * from './member';
-export * from './setting';
